Explosion

+++ Projet S3/Assets/Script/Player/KillCountPlayer.cs
@@ -8,7 +8,7 @@ public class KillCountPlayer : MonoBehaviour
     public int[] arrayOfKill = new int[0];
     public static List <float> killCount = new List<float>();
     public float timerOfKillCount = 5;
-    public int count;
+    public float count;
     private static float timerOfKill;
     public Text killCountUI;
